data-enrichment/requirements.txt
asttokens==3.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
cryptography==46.0.7
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
Flask==3.1.3
flask-cors==6.0.2
geographiclib==2.1
google-auth==2.49.1
google-auth-oauthlib==1.3.1
googlemaps==4.10.0
gspread==6.2.1
gspread-dataframe==4.0.0
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
packaging==26.0
pandas==3.0.2
parso==0.8.6
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.8
widgetsnbextension==4.0.15
python-dotenv==1.2.2
